requirements.txt
altgraph==0.17
bcrypt==3.2.0
beautifulsoup4==4.9.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
et-xmlfile==1.0.1
future==0.18.2
idna==2.10
jdcal==1.4.1
mysql==0.0.2
mysql-connector-python==8.0.21
mysqlclient==2.0.1
openpyxl==3.0.5
paramiko==2.7.2
pefile==2019.4.18
Pillow==7.2.0
protobuf==3.13.0
pycparser==2.20
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
PyNaCl==1.4.0
pywin32-ctypes==0.2.0
requests==2.24.0
scp==0.13.2
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
XlsxWriter==1.3.6
